fix: Allow php-fpm to run as root (#32)
The container should run as `nobody` (as is specified within the Dockerfile) but it's possible that a use-case may require the container run as root. The decision of which user to run as should not be blocked by php-fpm: without `--allow-to-run-as-root` it is not possible to use this image when running as root.